requirements.txt
flask==1.1.1
keras==2.1.5
numpy==1.17.4
opencv-python==4.1.1.26
pillow==6.2.0
PyQt5==5.13.2
tensorflow==1.8.0
matplotlib==3.1.1